+++ CHANGES.rst
@@ -7,6 +7,7 @@ Products.MeetingCharleroi Changelog
 - Be defensive when using getProperty on a member object, do not fail if member is None
 - Category 'indeterminee' can not be used on MeetingItemCollege if not to send to 'meting-config-council'
 - Added possibility to send and item that is 'prevalidated' back to 'proposed' and 'itemcreated'
+- Only a real Manager may backTo_prevalidated_from_waiting_advices
 
 4.0 (2017-08-22)
 ----------------